feat: add createErrorDialogApolloLink
