+++ packages/admin/src/error/errordialog/index.ts
@@ -3,3 +3,4 @@ export * from "./ErrorDialog";
 export * from "./ErrorDialogProvider";
 export * from "./useErrorDialog";
 export * from "./ErrorScope";
+export * from "./createErrorDialogApolloLink";
